Move helpers to separate lib
